# agent-workspace/04-technical-design/api-contracts/openapi.yaml
openapi: 3.0.0
info:
  title: 分行客户管理系统API
  description: 分行客户管理系统的RESTful API定义
  version: 1.0.0
  contact:
    name: 技术团队
    email: tech@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: http://dev.example.com/api/v1
    description: 开发环境
  - url: http://prod.example.com/api/v1
    description: 生产环境
paths:
...
  /visits:
    get:
      summary: 获取拜访记录列表
      description: 根据条件查询客户拜访记录
      tags:
        - 客户拜访管理
      parameters:
        - name: customerName
          in: query
          description: 客户名称
          schema:
            type: string
        - name: startDate
          in: query
          description: 开始日期
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: 结束日期
          schema:
            type: string
            format: date
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: 每页大小
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitListResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: 创建拜访记录
      description: 创建新的客户拜访记录
      tags:
        - 客户拜访管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /visits/{id}:
    get:
      summary: 获取拜访记录详情
      description: 根据ID获取客户拜访记录详情
      tags:
        - 客户拜访管理
      parameters:
        - name: id
          in: path
          description: 拜访记录ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '404':
          description: 拜访记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: 更新拜访记录
      description: 更新客户拜访记录
      tags:
        - 客户拜访管理
      parameters:
        - name: id
          in: path
          description: 拜访记录ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 拜访记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: 删除拜访记录
      description: 删除客户拜访记录
      tags:
        - 客户拜访管理
      parameters:
        - name: id
          in: path
          description: 拜访记录ID
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 删除成功
        '404':
          description: 拜访记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 礼品管理服务API
...
  /stats/visit:
    get:
      summary: 获取客户拜访统计数据
      description: 获取客户拜访的统计数据，支持按时间维度查询
      tags:
        - 数据统计
      parameters:
        - name: timeDimension
          in: query
          description: 时间维度（day, week, month, quarter, year）
          required: true
          schema:
            type: string
            enum: [day, week, month, quarter, year]
        - name: startDate
          in: query
          description: 开始日期
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: 结束日期
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitStatsResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stats/gift:
    get:
      summary: 获取礼品申请统计数据
      description: 获取礼品申请的统计数据，支持按时间维度查询
      tags:
        - 数据统计
      parameters:
        - name: timeDimension
          in: query
          description: 时间维度（day, week, month, quarter, year）
          required: true
          schema:
            type: string
            enum: [day, week, month, quarter, year]
        - name: startDate
          in: query
          description: 开始日期
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: 结束日期
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftStatsResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # AI集成服务API
...
components:
  schemas:
...
    VisitRequest:
      type: object
      required:
        - customerName
        - visitDate
        - visitContent
      properties:
        customerName:
          type: string
          description: 客户名称
        visitDate:
          type: string
          format: date-time
          description: 拜访日期时间
        visitContent:
          type: string
          description: 拜访内容
        customerContact:
          type: string
          description: 客户联系方式
        nextVisitPlan:
          type: string
          description: 下次拜访计划
    VisitResponse:
      type: object
      properties:
        id:
          type: integer
          description: 拜访记录ID
        customerName:
          type: string
          description: 客户名称
        visitDate:
          type: string
          format: date-time
          description: 拜访日期时间
        visitContent:
          type: string
          description: 拜访内容
        customerContact:
          type: string
          description: 客户联系方式
        nextVisitPlan:
          type: string
          description: 下次拜访计划
        createdBy:
          type: string
          description: 创建人
        createdTime:
          type: string
          format: date-time
          description: 创建时间
        updatedBy:
          type: string
          description: 更新人
        updatedTime:
          type: string
          format: date-time
          description: 更新时间
    VisitListResponse:
      type: object
      properties:
        total:
          type: integer
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页大小
        data:
          type: array
          items:
            $ref: '#/components/schemas/VisitResponse'

    # 礼品管理相关
...
    VisitStatsResponse:
      type: object
      properties:
        timeDimension:
          type: string
          description: 时间维度
        startDate:
          type: string
          format: date
          description: 开始日期
        endDate:
          type: string
          format: date
          description: 结束日期
        data:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
                description: 时间点
              count:
                type: integer
                description: 拜访次数
    GiftStatsResponse:
      type: object
      properties:
        timeDimension:
          type: string
          description: 时间维度
        startDate:
          type: string
          format: date
          description: 开始日期
        endDate:
          type: string
          format: date
          description: 结束日期
        data:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
                description: 时间点
              total:
                type: integer
                description: 申请总数
              approved:
                type: integer
                description: 审批通过数
              rejected:
                type: integer
                description: 审批拒绝数

    # AI聊天相关
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误代码
        message:
          type: string
          description: 错误信息
        details:
          type: string
          description: 错误详情
          nullable: true
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: 会话Cookie
security:
  - SessionAuth: []